requirements.txt
certifi==2020.6.20
cycler==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
patsy==0.5.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
statsmodels==0.12.0
wincertstore==0.2
